refactor: move match quality utils to own file
